requirements.txt
streamlit
nomic
chromadb
ollama
pypdf
pdf2image
pillow
python-dotenv
tiktoken
transformers
torch
torchvision
accelerate
einops
timm
